Containerized Database, and made it into postgreSQL instead of SQLite

+++ Surtr/main.py
@@ -1,7 +1,9 @@
 import sys
+from dotenv import load_dotenv
+import os
 from pathlib import Path
 import multiprocessing
-import sqlite3
+import psycopg2
 from base64 import b64encode
 import argparse
 import random
@@ -23,14 +25,34 @@ from util import (
 from subroutines import deserialize_ep
 from parties import Teller
 from VCaster import VCaster
+
+
+load_dotenv("../Database/.env")
 # pylint: disable=no-member
 
+#password = Path("../Database/password.txt").read_text().strip()
 # Set up the connection to the database
-con = sqlite3.connect("../Database/BulletinBoard.db")
+con = psycopg2.connect(
+    host="localhost",
+    port=5433,
+    dbname=os.getenv("DB_NAME"),
+    user=os.getenv("DB_USER"),
+    password=os.getenv("DB_PASSWORD"),
+)
 cur = con.cursor()
 
 # Starts each run by clearing the database
-cur.execute("DELETE FROM encryptedVotes")
+cur.execute("""
+    SELECT EXISTS (
+        SELECT FROM information_schema.tables 
+        WHERE table_schema = 'public' 
+        AND table_name = 'encryptedvotes'
+    )
+""")
+
+if cur.fetchone()[0]:
+    cur.execute("DELETE FROM encryptedVotes")
+    con.commit()
 
 # Decodes from x, y coordinates to EccPoints
 def decode_point_recursive(obj):
@@ -222,7 +244,8 @@ def voting():
         t_voting_single = t_voting_single + t_voting
 
 def retrieve_ballot(id):
-    ballot = cur.execute("SELECT * FROM encryptedVotes WHERE id = ?", (id,)).fetchone()
+    cur.execute("SELECT * FROM encryptedVotes WHERE id = %s", (id,))
+    ballot = cur.fetchone()
     return decode_bb_data(ballot)
 
 def tallying():
